+++ server/src/main/java/org/opensearch/cluster/routing/OperationRouting.java
@@ -400,7 +400,8 @@ public class OperationRouting {
                 )) return indexShard.onlyNodeActiveInitializingShardsIt(localNodeId);
                 else {
                     throw new PreferenceBasedSearchNotAllowedException(
-                        "Preference based routing not allowed on weigh away node with strict weighted shard routing setting"
+                        "Preference based routing not allowed on weigh away node with strict weighted shard routing " +
+                            "enabled"
                     );
                 }
             case ONLY_NODES: